requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
rdkit-pypi==2024.3.5
deepchem==2.8.0
torch>=2.0.0
torch-geometric>=2.4.0
biopython==1.84
chemprop==2.0.0
requests>=2.31.0
httpx>=0.27.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
seaborn>=0.13.0
openai>=1.40.0
jinja2>=3.1.0
python-multipart>=0.0.9
aiofiles>=24.1.0
